Add browsable Targets and Datasets pages with server-side pagination
Users can now browse all targets and datasets without searching first.
Two new pages (/targets, /datasets) provide faceted browsing with
server-side pagination (10 per page), search, and CSV download.
Home page search now redirects to these pages with query pre-filled.
